requirements.txt
django >= 1.11.4
djangorestframework >= 3.6.3
